requirements.txt
jenkins-webapi
